open_flamingo_serve/requirements.txt
jina>=3.15.0
docarray>=0.30.0
transformers>=4.28.0
open-flamingo==0.0.2
open_clip_torch==2.16.0

